rename file to file convention, rename method
