--- lib/models/from_postgres/map/temp.geojson
@@ -1,1 +1,0 @@
-{"type": "FeatureCollection", "features": [{"type": "Feature", "geometry": {"type": "Point", "coordinates": [-76.748811, 18.005752]}, "properties": {"title": "IFLT", "associatedWith": "", "description": "Inter-Faculty Lecture Theater", "altName": ""}}, {"type": "Feature", "geometry": {"type": "Point", "coordinates": [-76.749251, 18.006033]}, "properties": {"title": "Center for Marine Sciences", "associatedWith": "Life Sciences", "description": "Center for Marine Sciences", "altName": ""}}, {"type": "Feature", "geometry": {"type": "Point", "coordinates": [-76.748826, 18.006095]}, "properties": {"title": "Department of Geography and Geology", "associatedWith": "Geography and Geology", "description": "Department of Geography and Geology", "altName": ""}}, {"type": "Feature", "geometry": {"type": "Point", "coordinates": [-76.748894, 18.005434]}, "properties": {"title": "Faculty of Science and Technology Assembly Point", "associatedWith": "", "description": "Faculty of Science and Technology Assembly Point", "altName": ""}}, {"type": "Feature", "geometry": {"type": "Polygon", "coordinates": [[[-76.749292, 18.005488], [-76.749006, 18.005622], [-76.74892, 18.005468], [-76.749292, 18.005488]]]}, "properties": {"title": "Physics Lab Building", "associatedWith": "Physics", "description": "Physics Lab Building", "altName": ""}}, {"type": "Feature", "geometry": {"type": "Polygon", "coordinates": [[[-76.749331, 18.006038], [-76.749368, 18.00592], [-76.74908, 18.00584], [-76.749052, 18.00596], [-76.749331, 18.006038]]]}, "properties": {"title": "Center for Marine Sciences Building", "associatedWith": "Life Sciences", "description": "Center for Marine Sciences Building", "altName": ""}}, {"type": "Feature", "geometry": {"type": "Polygon", "coordinates": [[[-76.749272, 18.006198], [-76.749307, 18.006078], [-76.748768, 18.005925], [-76.748748, 18.005972], [76.748751, 18.006018], [-76.748723, 18.005964], [-76.7487, 18.006036], [-76.749272, 18.006198]]]}, "properties": {"title": "Department of Geography and Geology Building", "associatedWith": "", "description": "Department of Geography and Geology Building", "altName": ""}}]}
