+++ lib/screens/scholarship_screen/local_widget/scholar_card.dart
